Merged PR 160: Updated Dependency Scanning and Container Scanning Documentation
Corrected and updated missing documentation in this folder.
I'm gong to recommend that we move the SonarQube and Fossa and Even The WhiteSource Dependency scanning documents alongside the SonarCloud and Rosslyn Analysers Docs in the SCA folder and rename this folder to Container Scanning. I will push these changes in a seperate PR though as ther are a number of reference links that will need to be updated simultaneously

Related work items: #247481